Improve random selection.
Maintain a context per mount mode containing the current speed and array
of spell IDs for that speed.
Select a random ID from the list.

Results in a more random selection of mounts when a lot of mounts are favorites.

The previous behavior was biased towards the last few favorite mounts per
mode/speed.